Make `irc_crawler.py` use argparse
It appears that a lot of modules use `getopt`. Python 2.7 introduces the
`argparse` module (and it can be installed via pip on older
distributions). This migrates `irc_crawler.py` over to `argparse` as a
demonstration. This give (IMHO) cleaner argument parsing, and really
nice (IMHO) help output.

David, feel free to reject this patch if you don't like it. No feelings
will be hurt.